+++ NEWS.md
@@ -1,3 +1,12 @@
+# rice 0.4.1
+* corrected a bug in C14topMC where errors were not calculated correctly.
+* caldist now always glues a postbomb curve to a prebomb if postbomb is not FALSE.
+* new functions 'push.normal' and 'push.gamma' to push a date to younger or older ages by adding/subtracting a normal resp. gamma distribution.
+* reservoir effect (deltaR, deltaSTD) options have been added to functions where this is relevant.
+* added more warnings to 'calibrate' for very old dates.
+* TODO: add mc way to include error estimate for muck function. Also check that errors are included correctly in contaminate and clean functions.
+
+
 # rice 0.4.0
 * the `contaminate` function now also produces a plot, and more details of the calculations.
 * the `decontaminate` function has been renamed to `clean` and has been updated with clearer messages; it now removes contamination to calculate the true/target age. A plot is made, and calculation details are provided.